+++ outlook_util.py
@@ -9,6 +9,7 @@ import os
 import time
 from datetime import timezone
 from utility import configure_logging
+from dateutil import tz
 
 configure_logging()
 
@@ -76,6 +77,11 @@ def get_outlook_datetime_formatted(event):
     if event['isAllDay'] == True:
         day = start_date.strftime("%a %b %-d")
     else:
+        # Convert start/end to local time
+        start_date = start_date.replace(tzinfo=tz.tzutc())
+        start_date = start_date.astimezone(tz.tzlocal())
+        end_date = end_date.replace(tzinfo=tz.tzutc())
+        end_date = end_date.astimezone(tz.tzlocal())
         if(start_date.date() == end_date.date()):
             start_formatted = start_date.strftime("%a %b %-d, %-I:%M %p")
             end_formatted = end_date.strftime("%-I:%M %p")
@@ -112,7 +118,7 @@ def main():
             events_data = get_outlook_calendar_events(cal["id"], now_iso, oneyearlater_iso, access_token)
 
             for event in events_data["value"]:
-                print("     ", event["subject"], ": ", event["start"]["dateTime"])
+                print("     ", event["subject"], ": ", get_outlook_datetime_formatted(event))
             print("============================================")